types for request and response
