+++ app/react/V2/api/paragraphExtractor/entities.ts
@@ -2,68 +2,67 @@
 import { IncomingHttpHeaders } from 'http';
 import api from 'app/utils/api';
 import { RequestParams } from 'app/utils/RequestParams';
-import {
-  PXEntityApiResponse,
-  PXEntityQuery,
-  PXEntityTable,
-  PXTable,
-} from 'app/V2/Routes/Settings/ParagraphExtraction/types';
+import { PXEntityTable, PXTable } from 'V2/Routes/Settings/ParagraphExtraction/types';
+import { PXEntityQuery, PXEntityRows } from 'V2/shared/ParagraphExtractionTypes';
 
-const dummyData = [
-  {
-    _id: '1',
-    title: 'John Smith',
-    document: 'doc_name.pdf',
-    languages: ['uk', 'es'],
-    templateId: '66fbe4f28542cc5545e05a46',
-    paragraphCount: 3,
-    status: 'NEW',
-  },
-  {
-    _id: '2',
-    title: 'Maria Garcia',
-    document: 'another doc.pdf',
-    languages: ['en', 'es'],
-    templateId: '66fbe4f28542cc5545e05a46',
-    paragraphCount: 5,
-    status: 'DONE',
-  },
-  {
-    _id: '3',
-    title: 'Pierre Dubois',
-    document: 'third_doc.pdf',
-    languages: ['fr'],
-    templateId: '66fbe4f28542cc5545e05a46',
-    paragraphCount: 2,
-    status: 'IN_QUEUE',
-  },
-  {
-    _id: '4',
-    title: 'Hans Mueller',
-    document: 'fourth_doc.pdf',
-    languages: ['en', 'fr'],
-    templateId: '66fbe4f28542cc5545e05a46',
-    paragraphCount: 4,
-    status: 'HAS_ERROR',
-  },
-  {
-    _id: '5',
-    title: 'Giulia Rossi',
-    document: 'fifth_doc.pdf',
-    languages: ['en', 'es'],
-    templateId: '66fbe4f28542cc5545e05a46',
-    paragraphCount: 6,
-    status: 'PROCESSING',
-  },
-] as PXEntityApiResponse[];
+// const dummyData = [
+//   {
+//     _id: '1',
+//     title: 'John Smith',
+//     document: 'doc_name.pdf',
+//     languages: ['uk', 'es'],
+//     templateId: '66fbe4f28542cc5545e05a46',
+//     paragraphCount: 3,
+//     status: 'NEW',
+//   },
+//   {
+//     _id: '2',
+//     title: 'Maria Garcia',
+//     document: 'another doc.pdf',
+//     languages: ['en', 'es'],
+//     templateId: '66fbe4f28542cc5545e05a46',
+//     paragraphCount: 5,
+//     status: 'DONE',
+//   },
+//   {
+//     _id: '3',
+//     title: 'Pierre Dubois',
+//     document: 'third_doc.pdf',
+//     languages: ['fr'],
+//     templateId: '66fbe4f28542cc5545e05a46',
+//     paragraphCount: 2,
+//     status: 'IN_QUEUE',
+//   },
+//   {
+//     _id: '4',
+//     title: 'Hans Mueller',
+//     document: 'fourth_doc.pdf',
+//     languages: ['en', 'fr'],
+//     templateId: '66fbe4f28542cc5545e05a46',
+//     paragraphCount: 4,
+//     status: 'HAS_ERROR',
+//   },
+//   {
+//     _id: '5',
+//     title: 'Giulia Rossi',
+//     document: 'fifth_doc.pdf',
+//     languages: ['en', 'es'],
+//     templateId: '66fbe4f28542cc5545e05a46',
+//     paragraphCount: 6,
+//     status: 'PROCESSING',
+//   },
+// ] as PXEntityApiResponse[];
 
-const get = async (parameters: PXEntityQuery, headers?: IncomingHttpHeaders) => {
+const get = async (
+  parameters: PXEntityQuery,
+  headers?: IncomingHttpHeaders
+): Promise<PXEntityRows> => {
   try {
     const requestParams = new RequestParams(parameters, headers);
     // const { json: response } = await api.get(apiEndpoint, requestParams);
     const id = requestParams.data?.filter?.extractorId;
-    return dummyData || id;
-    // return response;
+    // return dummyData || id;
+    return undefined;
   } catch (e) {
     return e;
   }
