+++ app/react/V2/shared/ParagraphExtractionTypes.ts
@@ -1,3 +1,5 @@
+type EntityStatus = 'NEW' | 'IN_QUEUE' | 'PROCESSING' | 'DONE' | 'HAS_ERROR';
+
 type Extractor = {
   _id: string;
   sourceTemplateId: string;
@@ -12,4 +14,38 @@ type Extractor = {
   };
 };
 
-export type { Extractor };
+type PXEntityQuery = {
+  filter: {
+    extractorId: string;
+    status?: string[];
+    languages?: string[];
+  };
+  page?: {
+    number: number;
+    size: number;
+  };
+};
+
+type PXEntityRows = {
+  rows: {
+    entity: {
+      _id: string;
+      sharedId: string;
+      title: string;
+      language: string;
+    };
+    status: {
+      _id: string;
+      status: EntityStatus;
+    };
+    availableFileLanguages: string[];
+    paragraphsCount: number;
+  };
+  page: {
+    number: number;
+    size: number;
+  };
+  totalRows: number;
+};
+
+export type { Extractor, PXEntityQuery, PXEntityRows, EntityStatus };
